Move peers list from PeerDb to PeerManager
PeerDb will be only used for address management.
