+++ p2p/src/peer_manager/tests/ban.rs
@@ -241,7 +241,7 @@ where
         )
         .await;
     assert_eq!(peer_manager.handle_result(Some(peer_id), res).await, Ok(()));
-    assert!(!peer_manager.peerdb.is_active_peer(&peer_id));
+    assert!(!peer_manager.is_active_peer(&peer_id));
 
     // invalid version
     let res = peer_manager
@@ -261,7 +261,7 @@ where
         )
         .await;
     assert_eq!(peer_manager.handle_result(Some(peer_id), res).await, Ok(()));
-    assert!(!peer_manager.peerdb.is_active_peer(&peer_id));
+    assert!(!peer_manager.is_active_peer(&peer_id));
 
     // valid connection
     let address = B::new();
@@ -283,7 +283,7 @@ where
         .await;
     assert!(res.is_ok());
     assert_eq!(peer_manager.handle_result(Some(peer_id), res).await, Ok(()));
-    assert!(peer_manager.peerdb.is_active_peer(&peer_id));
+    assert!(peer_manager.is_active_peer(&peer_id));
     assert!(!peer_manager.peerdb.is_address_banned(&address.as_bannable()).unwrap());
 }
 
@@ -345,7 +345,7 @@ where
         )
         .await;
     assert_eq!(peer_manager.handle_result(Some(peer_id), res).await, Ok(()));
-    assert!(!peer_manager.peerdb.is_active_peer(&peer_id));
+    assert!(!peer_manager.is_active_peer(&peer_id));
 
     // invalid version
     let res = peer_manager
@@ -364,7 +364,7 @@ where
         )
         .await;
     assert_eq!(peer_manager.handle_result(Some(peer_id), res).await, Ok(()));
-    assert!(!peer_manager.peerdb.is_active_peer(&peer_id));
+    assert!(!peer_manager.is_active_peer(&peer_id));
 
     // valid connection
     let address = B::new();
@@ -385,7 +385,7 @@ where
         .await;
     assert!(res.is_ok());
     assert_eq!(peer_manager.handle_result(Some(peer_id), res).await, Ok(()));
-    assert!(peer_manager.peerdb.is_active_peer(&peer_id));
+    assert!(peer_manager.is_active_peer(&peer_id));
     assert!(!peer_manager.peerdb.is_address_banned(&address.as_bannable()).unwrap());
 }
